+++ lib/CMakeLists.txt
@@ -13,6 +13,7 @@ add_subdirectory(hash)
 add_subdirectory(heap)
 add_subdirectory(mem_blocks)
 add_subdirectory(os)
+add_subdirectory(utils)
 add_subdirectory_ifdef(CONFIG_SMF smf)
 add_subdirectory_ifdef(CONFIG_OPENAMP              open-amp)
 add_subdirectory_ifdef(CONFIG_ACPI acpi)
