+++ lib/utils/bitarray.c
