+++ lib/utils/dec.c
