+++ lib/utils/hex.c
