+++ lib/utils/rb.c
