+++ lib/utils/winstream.c
