+++ tests/unit/base64/main.c
@@ -27,7 +27,7 @@
 #include <zephyr/sys/base64.h>
 #include <zephyr/ztest.h>
 
-#include "../../../lib/os/base64.c"
+#include "../../../lib/utils/base64.c"
 
 static const unsigned char base64_test_dec[64] = {
 	0x24, 0x48, 0x6E, 0x56, 0x87, 0x62, 0x5A, 0xBD,
