+++ tests/unit/util/CMakeLists.txt
@@ -4,4 +4,4 @@ cmake_minimum_required(VERSION 3.20.0)
 
 project(util)
 find_package(Zephyr COMPONENTS unittest REQUIRED HINTS $ENV{ZEPHYR_BASE})
-target_sources(testbinary PRIVATE main.c maincxx.cxx ${ZEPHYR_BASE}/lib/os/dec.c)
+target_sources(testbinary PRIVATE main.c maincxx.cxx ${ZEPHYR_BASE}/lib/utils/dec.c)
